webkit-efl/LayoutTests/platform/efl-tizen/fast/frames/frameElement-widthheight-expected.txt

ALERT: Frame size: 489x735
